refactor: move all the component file into exsiting component folder
