+++ components/landing-page/Hero.jsx
