+++ components/landing-page/HowItWorks.jsx
